Connect BE delete handler to router & minor fixes

+++ server/src/controllers/comments.controllers/delete.comments.controller.ts
@@ -1,10 +1,16 @@
 import { Request, Response, NextFunction } from 'express';
 import { ICommentModel } from '../../models/comment.model';
+import { IUserModel } from '../../models/user.model';
 import { deleteComment, deleteCommentFromPost } from '../../db.requests/delete.comments.requests';
 
-export const addComments = async (req: Request, res: Response, next: NextFunction): Promise<void> => {
+export const deleteComments = async (req: Request, res: Response, next: NextFunction): Promise<void> => {
     try {
-        const { body: { commentId } }: Request = req;
+        const { body: { commentId, authorId } }: Request = req;
+        const { locals: { user: { _id: loggedUserId } } }: Response = res;
+
+        if (authorId !== loggedUserId) {
+            throw new Error(`Unauthorized attempt to delete comment ${commentId}.`);
+        }
 
         const deletedComment = await deleteComment(commentId, next);
         if (!deletedComment) {
